chore: move `get_error_type`
