+++ backend/substrapp/tests/compute_tasks/test_errors.py
@@ -1,9 +1,5 @@
-import io
-
 import pytest
 
-from builder import exceptions as build_errors
-from orchestrator import failure_report_pb2
 from substrapp.compute_tasks import errors
 
 
@@ -15,15 +11,3 @@ class TestComputeTaskErrorType:
     def test_from_int(self, input_value: int, expected: errors.ComputeTaskErrorType):
         result = errors.ComputeTaskErrorType.from_int(input_value)
         assert result == expected
-
-
-@pytest.mark.parametrize(
-    ("exc", "expected"),
-    [
-        (build_errors.BuildError(logs="some build error"), failure_report_pb2.ERROR_TYPE_BUILD),
-        (errors.ExecutionError(logs=io.BytesIO()), failure_report_pb2.ERROR_TYPE_EXECUTION),
-        (Exception(), failure_report_pb2.ERROR_TYPE_INTERNAL),
-    ],
-)
-def test_get_error_type(exc: Exception, expected: str):
-    assert errors.get_error_type(exc) == expected
